requirements.txt
asgiref==3.2.10
Django==3.1.2
django-filter==2.4.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
djangorestframework==3.12.1
numpy==1.19.2
phonenumbers==8.12.11
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.4.1
